try to parse new metadata format
