+++ tools/storage/Cargo.toml
@@ -15,12 +15,9 @@ rust-version = { workspace = true }
 [dependencies]
 anyhow = { workspace = true }
 clap = { workspace = true }
+diem-backup-cli = { workspace = true }
 diem-db-tool = { workspace = true }
 diem-logger = { workspace = true }
 diem-push-metrics = { workspace = true }
+serde_json = { workspace = true }
 tokio = { workspace = true }
-
-
-[[bin]]
-name = "storage"
-path = "main.rs"
