+++ tools/storage/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod read_snapshot;
